# Apply tricks for 2d zonotope vertex enumeration