--- a/src/Interfaces/AbstractZonotope.jl
+++ b/src/Interfaces/AbstractZonotope.jl
@@ -421,7 +421,7 @@ function vertices_list(Z::AbstractZonotope{N};
     if p == 0
         return [c]
     end
-    if n == 2
+    if n == 2 & p > 2
       if same_sign(G)
         sorted_G = sortslices(G, dims=2,by=x->atan(x[2], x[1]))
         index = ones(p, 2*p)