--- a/src/Interfaces/AbstractZonotope.jl
+++ b/src/Interfaces/AbstractZonotope.jl
@@ -422,7 +422,7 @@ function vertices_list(Z::AbstractZonotope{N};
         return [c]
     end
     if n == 2 && p > 2 && apply_convex_hull
-      if sum(abs.(G)) == sum(G)
+      if sum(abs, G) == abs(sum(G))
         sorted_G = sortslices(G, dims=2,by=x->atan(x[2], x[1]))
         index = ones(p, 2*p)
         for i in 1:p