--- a/src/Interfaces/AbstractZonotope.jl
+++ b/src/Interfaces/AbstractZonotope.jl
@@ -425,7 +425,7 @@ function vertices_list(Z::AbstractZonotope{N};
         if sum(abs, G) == abs(sum(G))
             sorted_G = sortslices(G, dims=2, by=x->atan(x[2], x[1]))
             index = ones(N, p, 2*p)
-            for i in 1:p
+            @inbounds for i in 1:p
                 index[i, i+1:i+p-1] *= -1
             end
             index[:, 1] *= -1