when generators are all in first/third quadrant