@mforets 2020-06-25
Thanks for your contribution! 

Here are some comments to get started:

- As per disussion on gitter, seems that we can generally check  `all(G .> 0) | all(G .< 0)` very efficiently by fusing both checks in one loop over the elements of `G`.

- Could you add a small description of the algorithm? This can be added in the `### Notes` section of the docstring of this  function.

- Why is `zeros(N, n, 1)` always added? For instance, `Z = Zonotope([1.0, 1.0], 0.1 * Matrix(1.0I, 2, 2))` satisfies the assumption on the generators's sign but the origin is not a vertex.

- The call `cumsum(sortslices(G,dims=2,by=x->x[2]/x[1]), dims=2)[:, 1:size(G,2)-1]` allocates a new vector.