@mforets 2020-06-26
Thanks for the update. I think now i understand how the algorithm works. Also thanks for the link https://math.stackexchange.com/questions/3356418/how-can-i-generate-boundary-in-2d-space-from-sum-of-finite-vectors/3356460#3356460

---


 About the fourth point in my last comment, i think there are some performance improvements that we can consider. First, i think that we can sort the generators only once. There are also some temporaries that can be cached such as the sum over all columns. We can consider using the internal `right_turn` for the comparisons which i presume may be faster than comparing by the quotients. I'll extend these comments when i have some free time.

For reference below i post the results so far that shows the new method in this branch has much better scalability. Nice!
It does not perform so well on small order zonotopes though. But my guess is that addressing some of the suggestions in the previous paragraph will improve this. In the comparison i also tried using static arrays. There are no perf difference (the algorithms are not specialized for these..).

```julia
function same_sign(G)
    s = sign(first(G))
    for j in size(G, 2)
        for i in size(G, 1)
            s != sign(G[i, j]) && return false
        end
    end
    return true
end

# new function
function _vertices_list_1d_v0(Z)
    G = Z.generators
    c = Z.center
    @assert same_sign(G)

    aux = sum(G, dims=2)
    A = -aux
    B = 2*cumsum(sortslices(G,dims=2,by=x->x[2]/x[1]), dims=2)[:, 1:size(G,2)-1] .- aux
    C = 2*cumsum(sortslices(G,dims=2,by=x->x[1]/x[2]), dims=2) .- aux
    V = hcat(A, B, C) .+ c
    [V[:, i] for i in 1:size(V, 2)]
end
```

Results:

```julia
for p in [2, 4, 6, 8, 10]
    println("p = $p")
    Z = Zonotope(rand(2), rand(2, p))
    Zs = Zonotope(SVector{2}(Z.center), SMatrix{2,p}(Z.generators))
    
    println(" - old - ")
    @btime vertices_list($Z)
    @btime vertices_list($Zs)

    println(" - new - ")
    @btime _vertices_list_1d_v0($Z)
    @btime _vertices_list_1d_v0($Zs)
    println(" ")
end

p = 2
 - old - 
  2.474 μs (45 allocations: 2.91 KiB)
  2.421 μs (53 allocations: 2.84 KiB)
 - new - 
  8.274 μs (68 allocations: 4.05 KiB)
  11.780 μs (70 allocations: 4.00 KiB)
 
p = 4
 - old - 
  10.142 μs (162 allocations: 12.66 KiB)
  9.418 μs (194 allocations: 12.66 KiB)
 - new - 
  8.572 μs (72 allocations: 5.11 KiB)
  12.296 μs (74 allocations: 5.53 KiB)
 
p = 6
 - old - 
  40.495 μs (597 allocations: 55.31 KiB)
  39.151 μs (725 allocations: 56.31 KiB)
 - new - 
  8.890 μs (76 allocations: 6.20 KiB)
  12.815 μs (78 allocations: 7.28 KiB)
 
p = 8
 - old - 
  169.002 μs (2329 allocations: 249.17 KiB)
  160.239 μs (2841 allocations: 257.17 KiB)
 - new - 
  9.192 μs (80 allocations: 7.23 KiB)
  13.395 μs (84 allocations: 9.66 KiB)
 
p = 10
 - old - 
  729.468 μs (9243 allocations: 1.09 MiB)
  692.103 μs (11291 allocations: 1.14 MiB)
 - new - 
  9.590 μs (84 allocations: 8.33 KiB)
  13.908 μs (88 allocations: 12.06 KiB)
```