@yupbank 2020-06-26
I made some changes to the `sorting` twice part.

Now we have 
```julia
function same_sign(G)
    s = sign(first(G))
    for j in size(G, 2)
        for i in size(G, 1)
            s != sign(G[i, j]) && return false
        end
    end
    return true
end

# new function
function _vertices_list_1d_v0(Z)
    G = Z.generators
    c = Z.center
    @assert same_sign(G)

    aux = sum(G, dims=2)
    A = -aux
    B = 2*cumsum(sortslices(G,dims=2,by=x->x[2]/x[1]), dims=2)[:, 1:size(G,2)-1] .- aux
    C = 2*cumsum(sortslices(G,dims=2,by=x->x[1]/x[2]), dims=2) .- aux
    V = hcat(A, B, C) .+ c
    [V[:, i] for i in 1:size(V, 2)]
end

# new function
function _vertices_list_1d_v2(Z)
    G = Z.generators
    c = Z.center
    @assert same_sign(G)
    p = size(G, 2)

    sorted_G = sortslices(G, dims=2,by=x->atan(x[2], x[1]))
    index = ones(p, 2*p)
    for i in 1:p
        index[i, i+1:i+p-1] *= -1
    end
    index[:, 1] *= -1
    V = sorted_G * index .+ c
    return [V[:, i] for i in 1:2*p]
end
```

And the performance part
``` Julia 
for p in [2, 3, 4, 5, 6, 8, 10]
    println("p = $p")
    Z = Zonotope(rand(2), rand(2, p))
    Zs = Zonotope(SVector{2}(Z.center), SMatrix{2,p}(Z.generators))
    
    println(" - old - ")
    @btime vertices_list($Z)
    @btime vertices_list($Zs)

    println(" - new - ")
    @btime _vertices_list_1d_v0($Z)
    @btime _vertices_list_1d_v0($Zs)
    println(" ")
    
    println(" - v2 - ")
    @btime _vertices_list_1d_v2($Z)
    @btime _vertices_list_1d_v2($Zs)
    println(" ")
end
```

gives 

```
p = 2
 - old - 
  3.164 μs (45 allocations: 2.83 KiB)
  4.012 μs (53 allocations: 2.83 KiB)
 - new - 
  13.267 μs (93 allocations: 4.73 KiB)
  17.200 μs (77 allocations: 4.05 KiB)
 
 - v2 - 
  5.604 μs (42 allocations: 2.69 KiB)
  4.704 μs (38 allocations: 2.73 KiB)
 
p = 3
 - old - 
  6.374 μs (89 allocations: 6.14 KiB)
  7.758 μs (105 allocations: 6.14 KiB)
 - new - 
  13.745 μs (97 allocations: 5.36 KiB)
  18.372 μs (79 allocations: 4.83 KiB)
 
 - v2 - 
  6.120 μs (47 allocations: 3.38 KiB)
  5.096 μs (42 allocations: 3.50 KiB)
 
p = 4
 - old - 
  11.797 μs (164 allocations: 12.47 KiB)
  14.911 μs (196 allocations: 12.72 KiB)
 - new - 
  14.401 μs (101 allocations: 5.86 KiB)
  18.115 μs (81 allocations: 5.64 KiB)
 
 - v2 - 
  6.719 μs (52 allocations: 4.13 KiB)
  5.679 μs (46 allocations: 4.38 KiB)
 
p = 5
 - old - 
  24.082 μs (310 allocations: 25.84 KiB)
  30.022 μs (374 allocations: 26.34 KiB)
 - new - 
  14.792 μs (105 allocations: 6.48 KiB)
  18.684 μs (83 allocations: 6.55 KiB)
 
 - v2 - 
  6.816 μs (57 allocations: 4.92 KiB)
  5.949 μs (50 allocations: 5.31 KiB)
 
p = 6
 - old - 
  47.217 μs (599 allocations: 54.36 KiB)
  60.108 μs (727 allocations: 56.36 KiB)
 - new - 
  14.088 μs (109 allocations: 6.98 KiB)
  17.940 μs (87 allocations: 7.52 KiB)
 
 - v2 - 
  7.519 μs (62 allocations: 5.83 KiB)
  6.510 μs (54 allocations: 6.39 KiB)
 
p = 8
 - old - 
  193.140 μs (2331 allocations: 245.20 KiB)
  246.819 μs (2843 allocations: 257.20 KiB)
 - new - 
  14.363 μs (117 allocations: 8.11 KiB)
  19.336 μs (91 allocations: 9.67 KiB)
 
 - v2 - 
  8.488 μs (72 allocations: 7.81 KiB)
  7.570 μs (62 allocations: 8.86 KiB)
 
p = 10
 - old - 
  858.423 μs (9245 allocations: 1.08 MiB)
  1.058 ms (11293 allocations: 1.14 MiB)
 - new - 
  14.871 μs (125 allocations: 9.23 KiB)
  19.255 μs (95 allocations: 12.08 KiB)
 
 - v2 - 
  10.250 μs (82 allocations: 10.06 KiB)
  9.230 μs (70 allocations: 11.72 KiB)
```