@mforets 2020-06-26
nice :+1: this is in the right direction. 

did you try if it makes a difference to sort using the right hand side rule? i think it would be: `sorted_G = sortslices(G, dims=2,lt=right_turn)`.